Refactor: remove Events.socials_dates
...in favour of calling SocialsListings directly.

There's something not 100% right here: the configuration of passing the
next 14 days of dates maybe deserves to live somewhere in its own
object? But this feels OK for now: it's only duplicated between two
places: the map socials listing and the main listing.
